refactor : 페이지 폴더명 수정
