+++ src/pages/login/find_password/style.tsx
@@ -1,0 +1,13 @@
+import { css } from '@emotion/react'
+
+export const passwordResetForm = css`
+  margin-top: 1.7em;
+`
+
+export const footButtonWrapper = css`
+  position: fixed;
+  bottom: 4.4em;
+  left: 0;
+  right: 0;
+  padding: 0 20px;
+`
