--- src/pages/login/password_email/style.tsx
@@ -1,5 +1,0 @@
-import { css } from '@emotion/react'
-
-export const passwordResetForm = css`
-  margin-top: 1.7em;
-`
