+++ src/pages/sign_up/completion/index.tsx
@@ -1,4 +1,3 @@
-import FootButton from '../../../components/Common/FootButton'
 import Back from '../../../components/Common/Header/back'
 import { css } from '@emotion/react'
 import { Common } from '../../../styles/common'
@@ -14,7 +13,7 @@ const SignUpCompletion = () => {
         </h2>
         <img src="/images/signup/character_com.svg" alt="프로필 완성 페이지 캐릭터" />
       </div>
-      <FootButton firstText="시작하기" secondText="" className1="orangeBtn" className="oneBtn" />
+      {/* <FootButton firstText="시작하기" secondText="" className1="orangeBtn" className="oneBtn" /> */}
     </>
   )
 }
