+++ src/pages/sign_up/profile_upload/index.tsx
@@ -1,4 +1,3 @@
-import FootButton from '../../../components/Common/FootButton'
 import Back from '../../../components/Common/Header/back'
 import Title from '../../../components/Common/Title'
 
@@ -9,7 +8,7 @@ const SignUpProfileUpload = () => {
 
       <Title title="프로필 사진을 올려주세요!" className="loginMt" />
 
-      <FootButton firstText="올리기 싫어요" secondText="등록하기" className1="whiteBtn" />
+      {/* <FootButton firstText="올리기 싫어요" secondText="등록하기" className1="whiteBtn" /> */}
     </>
   )
 }
